Add new types to common
