+++ datafusion/proto-common/src/to_proto/mod.rs
@@ -21,9 +21,7 @@ use crate::{protobuf_common as protobuf, protobuf_common};
 use arrow::datatypes::{
     DataType, Field, IntervalUnit, Schema, SchemaRef, TimeUnit, UnionMode,
 };
-use datafusion_common::{
-    plan_datafusion_err, Column, DFSchema, DFSchemaRef, DataFusionError, ScalarValue,
-};
+use datafusion_common::{plan_datafusion_err, Column, DFSchema, DFSchemaRef, DataFusionError, ScalarValue, Constraints, Constraint};
 use std::sync::Arc;
 
 impl From<Column> for protobuf::Column {
@@ -302,5 +300,34 @@ where
 impl From<Error> for DataFusionError {
     fn from(e: Error) -> Self {
         plan_datafusion_err!("{}", e)
+    }
+}
+
+impl From<Constraints> for protobuf::Constraints {
+    fn from(value: Constraints) -> Self {
+        let constraints = value.into_iter().map(|item| item.into()).collect();
+        protobuf::Constraints { constraints }
+    }
+}
+
+impl From<Constraint> for protobuf::Constraint {
+    fn from(value: Constraint) -> Self {
+        let res = match value {
+            Constraint::PrimaryKey(indices) => {
+                let indices = indices.into_iter().map(|item| item as u64).collect();
+                protobuf::constraint::ConstraintMode::PrimaryKey(
+                    protobuf::PrimaryKeyConstraint { indices },
+                )
+            }
+            Constraint::Unique(indices) => {
+                let indices = indices.into_iter().map(|item| item as u64).collect();
+                protobuf::constraint::ConstraintMode::PrimaryKey(
+                    protobuf::PrimaryKeyConstraint { indices },
+                )
+            }
+        };
+        protobuf::Constraint {
+            constraint_mode: Some(res),
+        }
     }
 }
