+++ datafusion/proto/src/logical_plan/to_proto.rs
@@ -26,7 +26,7 @@ use arrow::{
     record_batch::RecordBatch,
 };
 
-use datafusion_common::{Constraint, Constraints, ScalarValue, TableReference};
+use datafusion_common::{ScalarValue, TableReference};
 use datafusion_expr::expr::{
     self, AggregateFunctionDefinition, Alias, Between, BinaryExpr, Cast, GroupingSet,
     InList, Like, Placeholder, ScalarFunction, Sort, Unnest,
@@ -1129,35 +1129,6 @@ impl From<JoinConstraint> for protobuf::JoinConstraint {
         match t {
             JoinConstraint::On => protobuf::JoinConstraint::On,
             JoinConstraint::Using => protobuf::JoinConstraint::Using,
-        }
-    }
-}
-
-impl From<Constraints> for protobuf::Constraints {
-    fn from(value: Constraints) -> Self {
-        let constraints = value.into_iter().map(|item| item.into()).collect();
-        protobuf::Constraints { constraints }
-    }
-}
-
-impl From<Constraint> for protobuf::Constraint {
-    fn from(value: Constraint) -> Self {
-        let res = match value {
-            Constraint::PrimaryKey(indices) => {
-                let indices = indices.into_iter().map(|item| item as u64).collect();
-                protobuf::constraint::ConstraintMode::PrimaryKey(
-                    protobuf::PrimaryKeyConstraint { indices },
-                )
-            }
-            Constraint::Unique(indices) => {
-                let indices = indices.into_iter().map(|item| item as u64).collect();
-                protobuf::constraint::ConstraintMode::PrimaryKey(
-                    protobuf::PrimaryKeyConstraint { indices },
-                )
-            }
-        };
-        protobuf::Constraint {
-            constraint_mode: Some(res),
         }
     }
 }
